explode memory
